+++ qualitycheck-20190115/alibabacloud_qualitycheck20190115/models/_get_agent_task_result_request.py
@@ -10,7 +10,7 @@ class GetAgentTaskResultRequest(DaraModel):
         base_me_agent_id: str = None,
         json_str: str = None,
     ):
-        # The ID of the business workspace.
+        # The ID of the business space.
         self.base_me_agent_id = base_me_agent_id
         # The complete JSON string. For more information, see the following details.
         self.json_str = json_str